pyproject.toml
[tool.poetry]
name = "source-assignment-software-engineering"
version = "0.1.0"
description = ""
authors = ["Jan Hein de Jong <janhein.dejong@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.97.0"
uvicorn = "^0.20.0"
email-validator = "^2.0.0.post2"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
mypy = "^0.991"
ruff = "^0.0.270"
pytest = "^7.3.1"
httpx = "^0.24.1"
freezegun = "^1.2.2"
lxml = "^4.9.2"
types-python-jose = "^3.3.4.7"

[tool.poetry.scripts]
source_weather_cli = "app.cli:main"

[tool.mypy]
exclude = "./data"

[tool.black]
extend-exclude = "data"

[tool.ruff]
extend-exclude = ["data"]
select = ["E", "F", "B", "I"]
extend-ignore = ["B008"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
